+++ lib/hairballs/themes/turboladen.rb
@@ -1,8 +1,11 @@
 require 'hairballs'
 
 Hairballs.add_theme(:turboladen) do |theme|
+  IRB.conf[:AUTO_INDENT] = true
+
   theme.libraries do |libs_to_require|
     libs_to_require += %w(
+      awesome_print
       irb/completion
       looksee
       colorize
@@ -10,19 +13,18 @@ Hairballs.add_theme(:turboladen) do |theme|
 
     libs_to_require +=
       case RUBY_PLATFORM
-      when /mswin32|mingw32/
-        %w(win32console)
-      when /darwin/
-        %w(terminal-notifier)
-      else
-        []
+      when /mswin32|mingw32/ then %w(win32console)
+      when /darwin/ then %w(terminal-notifier)
+      else []
       end
 
     libs_to_require
   end
 
   theme.prompt do |prompt|
-    preface = proc do |status = '  '|
+    status = '  '.freeze
+
+    preface = proc do
       if Hairballs.project_name
         "⟪#{Hairballs.project_name.light_blue}⟫#{status}%03n"
       else
@@ -30,11 +32,10 @@ Hairballs.add_theme(:turboladen) do |theme|
       end
     end
 
-    prompt.auto_indent = true
     prompt.normal = "#{preface.call}:%i> "
     prompt.continued_string = "#{preface.call('❊%l'.yellow)}:%i> "
     prompt.continued_statement = "#{preface.call('❊?'.yellow)}:%i> "
     prompt.indented_code = "#{preface.call('✚ '.yellow)}:%i> "
-    prompt.return_format = "➥ %s\n"
+    prompt.return_format = "➥ %s\n".freeze
   end
 end